fixed linux install (hopefully) and updated tests
