+++ nivlink/__init__.py
@@ -4,5 +4,5 @@ __version__ = '0.2'
 
 from .epochs import Epochs
 from .raw import Raw
-from .screen import ScreenInfo
+from .screen import Screen
 from .projects import epoching_fht
